fix syntax check on mac
I noticed that syntax error highlighting wasn't working.

When I tried debugging the Extension itself, it suggested a
file not found error:

```
Recevied stderr data event with
[chuck]: no such file: 'test.ck"'
```

The previous code wrapped the path in double quotes, then
passed it to `cp.spawn`. That means that it was equivalent
to passing a quoted value at the command line.

You can reproduce the problem outside of the extension like so:

```
$ chuck '"/Users/nathanleiby/test.ck"'
[chuck]: no such file: 'test.ck"'
```

Notice the trailing `"` in the no such file error.
It seems that `chuck` command isn't able to handle the additional quotes.

I suggest removing them.

---

I'm running chuck 1.5.5.2

```
$ chuck --version

chuck version: 1.5.5.2 (chai)
   macOS | 64-bit [universal binary]
   audio driver: CoreAudio
   http://chuck.cs.princeton.edu/
   http://chuck.stanford.edu/
```